backend/tournamentmanagement/requirements.txt
asgiref==3.8.1
Django==4.2.13
sqlparse==0.5.0
typing-extensions==4.12.1
djangorestframework==3.14
django-cors-headers==4.0.0
channels
channels-redis
psycopg2-binary